divided horizontal readers and vertical readers into their own classes
