+++ LString/LString.h
@@ -22,6 +22,8 @@ using namespace std;
 //                       a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p,  q, r, s, t, u, v, w, x, y,  z
 const int legend[26] = { 1, 3, 3, 2, 1, 4, 2, 4, 1, 8, 5, 1, 3, 1, 1, 3, 10, 1, 1, 1, 1, 4, 4, 8, 4, 10};
 
+enum Type{HORIZONTAL, VERTICAL, UNDEFINED};
+
 struct Letter {
     int x, y;
     int val;
@@ -178,8 +180,5 @@ public:
         return hash;
     }
 };
-
-
-
 
 #endif
