+++ ScrabbleReader/ScrabbleReader.cpp
@@ -1,0 +1,5 @@
+//
+// Created by misc1 on 3/3/2022.
+//
+
+#include "ScrabbleReader.h"
