+++ cmake-build-debug/CMakeFiles/Makefile2
@@ -74,12 +74,12 @@ clean: CMakeFiles/scrabble-solver.dir/clean
 CMakeFiles/scrabble-solver.dir/all:
 	$(MAKE) -f CMakeFiles/scrabble-solver.dir/build.make CMakeFiles/scrabble-solver.dir/depend
 	$(MAKE) -f CMakeFiles/scrabble-solver.dir/build.make CMakeFiles/scrabble-solver.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles --progress-num=1,2,3,4 "Built target scrabble-solver"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles --progress-num=1,2,3,4,5,6 "Built target scrabble-solver"
 .PHONY : CMakeFiles/scrabble-solver.dir/all
 
 # Build rule for subdir invocation for target.
 CMakeFiles/scrabble-solver.dir/rule: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start /mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles 4
+	$(CMAKE_COMMAND) -E cmake_progress_start /mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles 6
 	$(MAKE) -f CMakeFiles/Makefile2 CMakeFiles/scrabble-solver.dir/all
 	$(CMAKE_COMMAND) -E cmake_progress_start /mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles 0
 .PHONY : CMakeFiles/scrabble-solver.dir/rule
