+++ cmake-build-debug/CMakeFiles/scrabble-solver.dir/DependInfo.cmake
@@ -4,8 +4,10 @@ set(CMAKE_DEPENDS_LANGUAGES
   )
 # The set of files for implicit dependencies of each language:
 set(CMAKE_DEPENDS_CHECK_CXX
-  "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/BoardReader/BoardReader.cpp" "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles/scrabble-solver.dir/BoardReader/BoardReader.cpp.o"
+  "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/HorizontalBoardReader/HorizontalBoardReader.cpp" "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles/scrabble-solver.dir/HorizontalBoardReader/HorizontalBoardReader.cpp.o"
   "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/LString/LString.cpp" "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles/scrabble-solver.dir/LString/LString.cpp.o"
+  "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/ScrabbleReader/ScrabbleReader.cpp" "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles/scrabble-solver.dir/ScrabbleReader/ScrabbleReader.cpp.o"
+  "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/VerticalBoardReader/VerticalBoardReader.cpp" "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles/scrabble-solver.dir/VerticalBoardReader/VerticalBoardReader.cpp.o"
   "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/main.cpp" "/mnt/c/Users/misc1/CLionProjects/scrabble-solver/cmake-build-debug/CMakeFiles/scrabble-solver.dir/main.cpp.o"
   )
 set(CMAKE_CXX_COMPILER_ID "GNU")
