+++ cmake-build-debug/CMakeFiles/scrabble-solver.dir/depend.internal
@@ -1,14 +1,19 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 3.16
 
-CMakeFiles/scrabble-solver.dir/BoardReader/BoardReader.cpp.o
- /mnt/c/Users/misc1/CLionProjects/scrabble-solver/BoardReader/BoardReader.cpp
- /mnt/c/Users/misc1/CLionProjects/scrabble-solver/BoardReader/BoardReader.h
+CMakeFiles/scrabble-solver.dir/HorizontalBoardReader/HorizontalBoardReader.cpp.o
+ /mnt/c/Users/misc1/CLionProjects/scrabble-solver/HorizontalBoardReader/HorizontalBoardReader.cpp
+ /mnt/c/Users/misc1/CLionProjects/scrabble-solver/HorizontalBoardReader/HorizontalBoardReader.h
  /mnt/c/Users/misc1/CLionProjects/scrabble-solver/LString/LString.h
 CMakeFiles/scrabble-solver.dir/LString/LString.cpp.o
  /mnt/c/Users/misc1/CLionProjects/scrabble-solver/LString/LString.cpp
  /mnt/c/Users/misc1/CLionProjects/scrabble-solver/LString/LString.h
-CMakeFiles/scrabble-solver.dir/main.cpp.o
- /mnt/c/Users/misc1/CLionProjects/scrabble-solver/BoardReader/BoardReader.h
+CMakeFiles/scrabble-solver.dir/VerticalBoardReader/VerticalBoardReader.cpp.o
  /mnt/c/Users/misc1/CLionProjects/scrabble-solver/LString/LString.h
+ /mnt/c/Users/misc1/CLionProjects/scrabble-solver/VerticalBoardReader/VerticalBoardReader.cpp
+ /mnt/c/Users/misc1/CLionProjects/scrabble-solver/VerticalBoardReader/VerticalBoardReader.h
+CMakeFiles/scrabble-solver.dir/main.cpp.o
+ /mnt/c/Users/misc1/CLionProjects/scrabble-solver/HorizontalBoardReader/HorizontalBoardReader.h
+ /mnt/c/Users/misc1/CLionProjects/scrabble-solver/LString/LString.h
+ /mnt/c/Users/misc1/CLionProjects/scrabble-solver/VerticalBoardReader/VerticalBoardReader.h
  /mnt/c/Users/misc1/CLionProjects/scrabble-solver/main.cpp
