+++ cmake-build-debug/CMakeFiles/scrabble-solver.dir/depend.make
@@ -1,14 +1,19 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 3.16
 
-CMakeFiles/scrabble-solver.dir/BoardReader/BoardReader.cpp.o: ../BoardReader/BoardReader.cpp
-CMakeFiles/scrabble-solver.dir/BoardReader/BoardReader.cpp.o: ../BoardReader/BoardReader.h
-CMakeFiles/scrabble-solver.dir/BoardReader/BoardReader.cpp.o: ../LString/LString.h
+CMakeFiles/scrabble-solver.dir/HorizontalBoardReader/HorizontalBoardReader.cpp.o: ../HorizontalBoardReader/HorizontalBoardReader.cpp
+CMakeFiles/scrabble-solver.dir/HorizontalBoardReader/HorizontalBoardReader.cpp.o: ../HorizontalBoardReader/HorizontalBoardReader.h
+CMakeFiles/scrabble-solver.dir/HorizontalBoardReader/HorizontalBoardReader.cpp.o: ../LString/LString.h
 
 CMakeFiles/scrabble-solver.dir/LString/LString.cpp.o: ../LString/LString.cpp
 CMakeFiles/scrabble-solver.dir/LString/LString.cpp.o: ../LString/LString.h
 
-CMakeFiles/scrabble-solver.dir/main.cpp.o: ../BoardReader/BoardReader.h
+CMakeFiles/scrabble-solver.dir/VerticalBoardReader/VerticalBoardReader.cpp.o: ../LString/LString.h
+CMakeFiles/scrabble-solver.dir/VerticalBoardReader/VerticalBoardReader.cpp.o: ../VerticalBoardReader/VerticalBoardReader.cpp
+CMakeFiles/scrabble-solver.dir/VerticalBoardReader/VerticalBoardReader.cpp.o: ../VerticalBoardReader/VerticalBoardReader.h
+
+CMakeFiles/scrabble-solver.dir/main.cpp.o: ../HorizontalBoardReader/HorizontalBoardReader.h
 CMakeFiles/scrabble-solver.dir/main.cpp.o: ../LString/LString.h
+CMakeFiles/scrabble-solver.dir/main.cpp.o: ../VerticalBoardReader/VerticalBoardReader.h
 CMakeFiles/scrabble-solver.dir/main.cpp.o: ../main.cpp
