+++ cmake-build-debug/CMakeFiles/scrabble-solver.dir/link.txt
@@ -1,1 +1,1 @@
-/usr/bin/c++  -g  -rdynamic CMakeFiles/scrabble-solver.dir/main.cpp.o CMakeFiles/scrabble-solver.dir/BoardReader/BoardReader.cpp.o CMakeFiles/scrabble-solver.dir/LString/LString.cpp.o  -o scrabble-solver 
+/usr/bin/c++  -g  -rdynamic CMakeFiles/scrabble-solver.dir/main.cpp.o CMakeFiles/scrabble-solver.dir/HorizontalBoardReader/HorizontalBoardReader.cpp.o CMakeFiles/scrabble-solver.dir/LString/LString.cpp.o CMakeFiles/scrabble-solver.dir/VerticalBoardReader/VerticalBoardReader.cpp.o CMakeFiles/scrabble-solver.dir/ScrabbleReader/ScrabbleReader.cpp.o  -o scrabble-solver
