+++ cmake-build-debug/CMakeFiles/scrabble-solver.dir/progress.make
@@ -2,4 +2,6 @@ CMAKE_PROGRESS_1 = 1
 CMAKE_PROGRESS_2 = 2
 CMAKE_PROGRESS_3 = 3
 CMAKE_PROGRESS_4 = 4
+CMAKE_PROGRESS_5 = 5
+CMAKE_PROGRESS_6 = 6
