+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Mar 03 17:52 CST
+Start testing: Mar 03 20:05 CST
 ----------------------------------------------------------
-End testing: Mar 03 17:52 CST
+End testing: Mar 03 20:05 CST
